movie_back/requirements.txt
# Flask 后端框架
Flask==3.0.0
flask-cors==4.0.0

# AI 和机器学习
sentence-transformers==2.7.0
chromadb==0.4.22
openai>=1.30.0
torch>=2.6.0
torchvision>=0.21.0
numpy==1.24.3
scipy==1.11.1
huggingface-hub>=0.20.0

# BM25 算法和中文分词
rank-bm25==0.2.2
jieba==0.42.1

# Qwen Rerank
dashscope==1.20.5

# 数据处理
pandas==2.0.3

# 环境变量管理
python-dotenv==1.0.0

# HTTP 请求
requests==2.31.0

# 认证和安全
pyjwt==2.8.0
bcrypt==4.1.2

# 工具
colorama==0.4.4
